Seperate module for user config object added
