+++ uaclient/cli/tests/test_cli_config_set.py
@@ -83,8 +83,7 @@ class TestMainConfigSet:
         assert err_msg in err
 
 
-@mock.patch("uaclient.config.state_files.user_config_file_private.write")
-@mock.patch("uaclient.config.state_files.user_config_file_public.write")
+@mock.patch("uaclient.config.user_config_file.user_config.write")
 @mock.patch("uaclient.cli.contract.get_available_resources")
 class TestActionConfigSet:
     @mock.patch("uaclient.util.we_are_currently_root", return_value=False)
@@ -92,8 +91,7 @@ class TestActionConfigSet:
         self,
         _m_resources,
         _we_are_currently_root,
-        _write_public,
-        _write_private,
+        _write_,
         FakeConfig,
     ):
         """Root is required to run pro config set."""
@@ -122,8 +120,7 @@ class TestActionConfigSet:
         livepatch_status,
         configure_livepatch_proxy,
         _m_resources,
-        _write_public,
-        _write_private,
+        _write,
         key,
         value,
         livepatch_enabled,
@@ -186,8 +183,7 @@ class TestActionConfigSet:
         validate_proxy,
         configure_apt_proxy,
         _m_resources,
-        _write_public,
-        _write_private,
+        _write,
         key,
         value,
         scope,
@@ -294,8 +290,7 @@ class TestActionConfigSet:
         validate_proxy,
         configure_apt_proxy,
         _m_resources,
-        _write_public,
-        _write_private,
+        _write,
         key,
         value,
         scope,
@@ -405,8 +400,7 @@ class TestActionConfigSet:
         validate_proxy,
         configure_apt_proxy,
         _m_resources,
-        _write_public,
-        _write_private,
+        _write,
         key,
         value,
         scope,
@@ -469,8 +463,7 @@ class TestActionConfigSet:
         self,
         setup_apt_proxy,
         _m_resources,
-        _write_public,
-        _write_private,
+        _write,
         key,
         value,
         scope,
@@ -509,8 +502,7 @@ class TestActionConfigSet:
         self,
         setup_apt_proxy,
         _m_resources,
-        _write_public,
-        _write_private,
+        _write,
         key,
         value,
         scope,
@@ -528,9 +520,7 @@ class TestActionConfigSet:
         assert 1 == setup_apt_proxy.call_count
         assert [mock.call(**kwargs)] == setup_apt_proxy.call_args_list
 
-    def test_set_timer_interval(
-        self, _m_resources, _write_public, _write_private, FakeConfig
-    ):
+    def test_set_timer_interval(self, _m_resources, _write, FakeConfig):
         args = mock.MagicMock(key_value_pair="update_messaging_timer=28800")
         cfg = FakeConfig()
         action_config_set(args, cfg=cfg)
@@ -540,8 +530,7 @@ class TestActionConfigSet:
     def test_error_when_interval_is_not_valid(
         self,
         _m_resources,
-        _write_public,
-        _write_private,
+        _write,
         FakeConfig,
         invalid_value,
     ):
