+++ uaclient/config.py
@@ -28,7 +28,7 @@ from uaclient.defaults import (
     DEFAULT_DATA_DIR,
     PRIVATE_SUBDIR,
 )
-from uaclient.files import notices, state_files
+from uaclient.files import notices, user_config_file
 from uaclient.files.notices import Notice
 from uaclient.yaml import safe_load
 
@@ -39,18 +39,6 @@ MERGE_ID_KEY_MAP = {
     "resourceEntitlements": "type",
 }
 UNSET_SETTINGS_OVERRIDE_KEY = "_unset"
-
-# Proxy fields that are visible and configurable
-PROXY_FIELDS = [
-    "apt_http_proxy",
-    "apt_https_proxy",
-    "global_apt_http_proxy",
-    "global_apt_https_proxy",
-    "ua_apt_http_proxy",
-    "ua_apt_https_proxy",
-    "http_proxy",
-    "https_proxy",
-]
 
 # Keys visible and configurable using `pro config set|unset|show` subcommands
 UA_CONFIGURABLE_KEYS = (
@@ -109,7 +97,7 @@ class UAConfig:
     def __init__(
         self,
         cfg: Optional[Dict[str, Any]] = None,
-        user_config: Optional[state_files.UserConfigData] = None,
+        user_config: Optional[user_config_file.UserConfigData] = None,
         series: Optional[str] = None,
     ) -> None:
         """"""
@@ -126,17 +114,18 @@ class UAConfig:
         else:
             try:
                 self.user_config = (
-                    read_user_config() or state_files.UserConfigData()
+                    user_config_file.user_config.read()
+                    or user_config_file.UserConfigData()
                 )
             except Exception as e:
                 LOG.warning("Error loading user config", exc_info=e)
                 LOG.warning("Using default config values")
-                self.user_config = state_files.UserConfigData()
+                self.user_config = user_config_file.UserConfigData()
 
         # support old ua_config values in uaclient.conf as user-config.json
         # value overrides
         if "ua_config" in self.cfg:
-            self.user_config = state_files.UserConfigData.from_dict(
+            self.user_config = user_config_file.UserConfigData.from_dict(
                 {**self.user_config.to_dict(), **self.cfg["ua_config"]},
                 optional_type_errors_become_null=True,
             )
@@ -174,7 +163,7 @@ class UAConfig:
     @http_proxy.setter
     def http_proxy(self, value: str):
         self.user_config.http_proxy = value
-        write_user_config(self.user_config)
+        user_config_file.user_config.write(self.user_config)
 
     @property
     def https_proxy(self) -> Optional[str]:
@@ -183,7 +172,7 @@ class UAConfig:
     @https_proxy.setter
     def https_proxy(self, value: str):
         self.user_config.https_proxy = value
-        write_user_config(self.user_config)
+        user_config_file.user_config.write(self.user_config)
 
     @property
     def ua_apt_https_proxy(self) -> Optional[str]:
@@ -192,7 +181,7 @@ class UAConfig:
     @ua_apt_https_proxy.setter
     def ua_apt_https_proxy(self, value: str):
         self.user_config.ua_apt_https_proxy = value
-        write_user_config(self.user_config)
+        user_config_file.user_config.write(self.user_config)
 
     @property
     def ua_apt_http_proxy(self) -> Optional[str]:
@@ -201,7 +190,7 @@ class UAConfig:
     @ua_apt_http_proxy.setter
     def ua_apt_http_proxy(self, value: str):
         self.user_config.ua_apt_http_proxy = value
-        write_user_config(self.user_config)
+        user_config_file.user_config.write(self.user_config)
 
     @property  # type: ignore
     @lru_cache(maxsize=None)
@@ -225,7 +214,7 @@ class UAConfig:
         self.user_config.global_apt_http_proxy = value
         self.user_config.apt_http_proxy = None
         UAConfig.global_apt_http_proxy.fget.cache_clear()  # type: ignore
-        write_user_config(self.user_config)
+        user_config_file.user_config.write(self.user_config)
 
     @property  # type: ignore
     @lru_cache(maxsize=None)
@@ -249,7 +238,7 @@ class UAConfig:
         self.user_config.global_apt_https_proxy = value
         self.user_config.apt_https_proxy = None
         UAConfig.global_apt_https_proxy.fget.cache_clear()  # type: ignore
-        write_user_config(self.user_config)
+        user_config_file.user_config.write(self.user_config)
 
     @property
     def update_messaging_timer(self) -> int:
@@ -261,7 +250,7 @@ class UAConfig:
     @update_messaging_timer.setter
     def update_messaging_timer(self, value: int):
         self.user_config.update_messaging_timer = value
-        write_user_config(self.user_config)
+        user_config_file.user_config.write(self.user_config)
 
     @property
     def metering_timer(self) -> int:
@@ -273,7 +262,7 @@ class UAConfig:
     @metering_timer.setter
     def metering_timer(self, value: int):
         self.user_config.metering_timer = value
-        write_user_config(self.user_config)
+        user_config_file.user_config.write(self.user_config)
 
     @property
     def poll_for_pro_license(self) -> bool:
@@ -288,7 +277,7 @@ class UAConfig:
     @poll_for_pro_license.setter
     def poll_for_pro_license(self, value: bool):
         self.user_config.poll_for_pro_license = value
-        write_user_config(self.user_config)
+        user_config_file.user_config.write(self.user_config)
 
     @property
     def polling_error_retry_delay(self) -> int:
@@ -302,7 +291,7 @@ class UAConfig:
     @polling_error_retry_delay.setter
     def polling_error_retry_delay(self, value: int):
         self.user_config.polling_error_retry_delay = value
-        write_user_config(self.user_config)
+        user_config_file.user_config.write(self.user_config)
 
     @property
     def apt_news(self) -> bool:
@@ -314,7 +303,7 @@ class UAConfig:
     @apt_news.setter
     def apt_news(self, value: bool):
         self.user_config.apt_news = value
-        write_user_config(self.user_config)
+        user_config_file.user_config.write(self.user_config)
 
     @property
     def apt_news_url(self) -> str:
@@ -326,7 +315,7 @@ class UAConfig:
     @apt_news_url.setter
     def apt_news_url(self, value: str):
         self.user_config.apt_news_url = value
-        write_user_config(self.user_config)
+        user_config_file.user_config.write(self.user_config)
 
     def check_lock_info(self) -> Tuple[int, str]:
         """Return lock info if config lock file is present the lock is active.
@@ -709,38 +698,3 @@ def apply_config_settings_override(override_key: str):
         return new_f
 
     return wrapper
-
-
-def redact_config_data(
-    user_config: Optional[state_files.UserConfigData],
-) -> Optional[state_files.UserConfigData]:
-    """Redact the user-config.json file and return its contents."""
-    if user_config:
-        redacted_data = user_config
-        for field in PROXY_FIELDS:
-            value = getattr(redacted_data, field)
-            if value:
-                setattr(
-                    redacted_data, field, util.redact_sensitive_proxy(value)
-                )
-        return redacted_data
-    return None
-
-
-def read_user_config() -> Optional[state_files.UserConfigData]:
-    """Read the user-config.json file and return its contents."""
-    try:
-        if util.we_are_currently_root():
-            return state_files.user_config_file_private.read()
-        public_config = state_files.user_config_file_public.read()
-        return redact_config_data(public_config)
-    except Exception as e:
-        LOG.warning("Error loading user config", exc_info=e)
-        LOG.warning("Using default config values")
-        return None
-
-
-def write_user_config(user_config: state_files.UserConfigData) -> None:
-    """Write the user-config.json file with the given contents."""
-    state_files.user_config_file_public.write(user_config)
-    state_files.user_config_file_private.write(user_config)
