+++ uaclient/files/__init__.py
@@ -1,9 +1,15 @@
 from uaclient.files.data_types import DataObjectFile, DataObjectFileFormat
 from uaclient.files.files import MachineTokenFile, UAFile
+from uaclient.files.user_config_file import (
+    UserConfigData,
+    UserConfigFileObject,
+)
 
 __all__ = [
     "DataObjectFile",
     "DataObjectFileFormat",
     "MachineTokenFile",
     "UAFile",
+    "UserConfigData",
+    "UserConfigFileObject",
 ]
