+++ uaclient/files/state_files.py
@@ -1,5 +1,4 @@
 import datetime
-import os
 from typing import Any, Dict, List, Optional
 
 from uaclient import defaults
@@ -174,77 +173,6 @@ livepatch_support_cache = DataObjectFile(
     UserCacheFile("livepatch-kernel-support-cache.json"),
     file_format=DataObjectFileFormat.JSON,
 )
-
-
-class UserConfigData(DataObject):
-    fields = [
-        Field("apt_http_proxy", StringDataValue, required=False),
-        Field("apt_https_proxy", StringDataValue, required=False),
-        Field("global_apt_http_proxy", StringDataValue, required=False),
-        Field("global_apt_https_proxy", StringDataValue, required=False),
-        Field("ua_apt_http_proxy", StringDataValue, required=False),
-        Field("ua_apt_https_proxy", StringDataValue, required=False),
-        Field("http_proxy", StringDataValue, required=False),
-        Field("https_proxy", StringDataValue, required=False),
-        Field("apt_news", BoolDataValue, required=False),
-        Field("apt_news_url", StringDataValue, required=False),
-        Field("poll_for_pro_license", BoolDataValue, required=False),
-        Field("polling_error_retry_delay", IntDataValue, required=False),
-        Field("metering_timer", IntDataValue, required=False),
-        Field("update_messaging_timer", IntDataValue, required=False),
-    ]
-
-    def __init__(
-        self,
-        apt_http_proxy: Optional[str] = None,
-        apt_https_proxy: Optional[str] = None,
-        global_apt_http_proxy: Optional[str] = None,
-        global_apt_https_proxy: Optional[str] = None,
-        ua_apt_http_proxy: Optional[str] = None,
-        ua_apt_https_proxy: Optional[str] = None,
-        http_proxy: Optional[str] = None,
-        https_proxy: Optional[str] = None,
-        apt_news: Optional[bool] = None,
-        apt_news_url: Optional[str] = None,
-        poll_for_pro_license: Optional[bool] = None,
-        polling_error_retry_delay: Optional[int] = None,
-        metering_timer: Optional[int] = None,
-        update_messaging_timer: Optional[int] = None,
-    ):
-        self.apt_http_proxy = apt_http_proxy
-        self.apt_https_proxy = apt_https_proxy
-        self.global_apt_http_proxy = global_apt_http_proxy
-        self.global_apt_https_proxy = global_apt_https_proxy
-        self.ua_apt_http_proxy = ua_apt_http_proxy
-        self.ua_apt_https_proxy = ua_apt_https_proxy
-        self.http_proxy = http_proxy
-        self.https_proxy = https_proxy
-        self.apt_news = apt_news
-        self.apt_news_url = apt_news_url
-        self.poll_for_pro_license = poll_for_pro_license
-        self.polling_error_retry_delay = polling_error_retry_delay
-        self.metering_timer = metering_timer
-        self.update_messaging_timer = update_messaging_timer
-
-
-user_config_file_private = DataObjectFile(
-    UserConfigData,
-    UAFile(
-        "user-config.json",
-        os.path.join(defaults.DEFAULT_DATA_DIR, defaults.PRIVATE_SUBDIR),
-        private=True,
-    ),
-    DataObjectFileFormat.JSON,
-    optional_type_errors_become_null=True,
-)
-
-user_config_file_public = DataObjectFile(
-    UserConfigData,
-    UAFile("user-config.json", private=False),
-    DataObjectFileFormat.JSON,
-    optional_type_errors_become_null=True,
-)
-
 
 reboot_cmd_marker_file = UAFile("marker-reboot-cmds-required")
